requirements.txt
# Minimal deps for moderate_image.py
# (Some packages like torch/tensorflow are large and may take time to install.)
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
ImageHash>=4.3.1
numpy>=1.26.0
opennsfw2>=0.1.0
nudenet>=3.4.2
ultralytics>=8.0.0
pytesseract>=0.3.10
